Created GameRules folder and moved some validators there; started refactoring AvailableSpacesValidator class
