+++ lib/tic_tac_toe_rz/tictactoeruby.core/gamerules/tie_game_rules.rb
